Solutions all reformatted
