+++ Linked Lists/Cycle/ConstantSpace.java
@@ -1,0 +1,17 @@
+class Solution {
+  public boolean hasCycle(ListNode head) {
+    ListNode slow = head;
+    ListNode fast = head;
+
+    while (fast != null && fast.next != null) {
+      slow = slow.next;
+      fast = fast.next.next;
+
+      if (slow == fast) {
+        return true;
+      }
+    }
+
+    return false;
+  }
+}
