+++ Linked Lists/LinkedListFundamentals/PrintListInReverseRecursive.java
@@ -1,0 +1,21 @@
+class Solution {
+  public void printListInReverse(ListNode node) {
+    if (node == null) {
+      return;
+    }
+
+    /*
+      Recursion first - drill down to the end of the list. Base
+      case will catch us.
+    */
+    printListInReverse(node.next);
+
+    /*
+      Our work. This will be fired by the last node first, then
+      the 2nd to last node, and so on...because we recursed first,
+      got a ton of nodes on the stack in recursive calls, and now
+      they will run this as the recursion goes "back up"
+    */
+    System.out.println(node.value);
+  }
+}
