+++ Linked Lists/OddEvenPartition/UsingDummyHead.java
@@ -1,4 +1,4 @@
-class WithDummyHead {
+class Solution {
   public ListNode oddEvenList(ListNode head) {
     if (head == null || head.next == null) {
       return head;
@@ -29,28 +29,5 @@ class WithDummyHead {
     oddTail.next = null;
 
     return evenDummy.next;
-  }
-}
-
-class WithLessPointers {
-  public ListNode oddEvenList(ListNode head) {
-    if (head == null || head.next == null) {
-      return head;
-    }
-
-    ListNode even = head;
-    ListNode odd = head.next;
-
-    ListNode oddHead = odd;
-
-    while (odd != null && odd.next != null) {
-      even.next = odd.next;
-      even = odd.next;
-      odd.next = even.next;
-      odd = even.next;
-    }
-    even.next = oddHead;
-
-    return head;
   }
 }
